Use float variants of sqrt() and fabs().

+++ src/client/entities.c
@@ -93,9 +93,9 @@ entity_update_old(centity_t *ent, const entity_state_t *state, const vec_t *orig
         || state->modelindex4 != ent->current.modelindex4
         || event == EV_PLAYER_TELEPORT
         || event == EV_OTHER_TELEPORT
-        || fabs(origin[0] - ent->current.origin[0]) > 512
-        || fabs(origin[1] - ent->current.origin[1]) > 512
-        || fabs(origin[2] - ent->current.origin[2]) > 512
+        || fabsf(origin[0] - ent->current.origin[0]) > 512
+        || fabsf(origin[1] - ent->current.origin[1]) > 512
+        || fabsf(origin[2] - ent->current.origin[2]) > 512
         || cl_nolerp->integer == 1) {
         // some data changes will force no lerping
         ent->trailcount = 1024;     // for diminishing rocket / grenade trails
@@ -1068,7 +1068,7 @@ static void CL_SetupThirdPersionView(void)
     }
 
     VectorSubtract(focus, cl.refdef.vieworg, focus);
-    dist = sqrt(focus[0] * focus[0] + focus[1] * focus[1]);
+    dist = sqrtf(focus[0] * focus[0] + focus[1] * focus[1]);
 
     cl.refdef.viewangles[PITCH] = -RAD2DEG(atan2(focus[2], dist));
     cl.refdef.viewangles[YAW] -= cl_thirdperson_angle->value;
